Fixed problems with querying the database for posts and topics and displaying them
The queries were being set up but not executed. This code changes that. This code also fixed the code in the jade views to create links to the posts and topics in the database. This patch does too many things (I know, I realize that. I could break it up if needed.)

Signed-off-by: Lidza Louina <lidza.louina@gmail.com>