Adds asyncjs and reformats router params
This patch adds asyncjs to the code to get rid of callback hell.

It also reformats the way that data is passed to the router. Previously there was one variable for the database, but I realized that I'll also need helpers. And if I need more pieces of data passed into it, it would be a huge mess. So now it is one variable called routerData that contains a hash for the database and the helpers. And if I need more types of data for it, it will be easy to add.

Signed-off-by: Lidza Louina <lidza.louina@gmail.com>